requirements.txt
alabaster==0.7.10
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.0
attrs==17.3.0
autoflake==0.6.6
Automat==0.6.0
autopep8==1.3.3
Babel==2.5.1
bandit==1.4.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.6
cmakelint==1.3.4.1
coala==0.11.0
coala-bears==0.11.1
coala-utils==0.6.7
colorama==0.3.9
colorlog==2.10.0
constantly==15.1.0
cppclean==0.12
cpplint==1.3.0
cryptography==2.1.4
cssselect==1.0.1
decorator==4.1.2
dennis==0.9
dependency-management==0.4.0
Django==1.11.8
django-distill==1.2
docopt==0.6.2
docutils==0.14
eradicate==0.1.6
future==0.16.0
git-url-parse==1.0.2
gitdb==0.6.4
GitPython==1.0.1
guess-language-spirit==0.5.3
html-linter==0.3.0
hyperlink==17.3.1
idna==2.6
IGitt==0.4.1
imagesize==0.7.1
incremental==17.5.0
ipython-genutils==0.2.0
isort==4.2.15
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
libclang-py3==3.4.0
lxml==4.1.1
mando==0.3.3
markdown2==2.3.5
MarkupSafe==1.0
mccabe==0.6.1
munkres3==1.0.5.5
mypy-lang==0.4.6
nbformat==4.4.0
nltk==3.2.5
olefile==0.44
packaging==16.8
parsel==1.2.0
pbr==3.1.1
Pillow==4.3.0
polib==1.1.0
proselint==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
PyDispatcher==2.0.5
pydocstyle==1.1.1
pyflakes==1.4.0
Pygments==2.2.0
pylint==1.8.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPrint==0.2.6
pyroma==2.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
queuelib==1.4.2
radon==1.4.0
requests==2.18.4
restructuredtext-lint==1.0.1
rstcheck==2.2
ruamel.yaml==0.15.35
safety==0.5.1
sarge==0.1.4
Scrapy==1.4.0
scspell3k==2.1
service-identity==17.0.0
six==1.11.0
smmap==0.9.0
snowballstemmer==1.2.1
Sphinx==1.4.9
stevedore==1.28.0
template-remover==0.1.9
termcolor==1.1.0
testfixtures==4.14.3
traitlets==4.3.2
Twisted==17.9.0
unidiff==0.5.4
urllib3==1.22
vulture==0.10
w3lib==1.18.0
wrapt==1.10.11
yamllint==1.6.1
yapf==0.14.0
zope.interface==4.4.3
